+++ app/src/main/java/com/finance_tracker/finance_tracker/screens/main/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.finance_tracker.finance_tracker.sreens.operations
+package com.finance_tracker.finance_tracker.screens.main
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -14,30 +14,28 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import com.finance_tracker.finance_tracker.R
-import com.finance_tracker.finance_tracker.sreens.transactions.TransactionsListScreen
 import com.finance_tracker.finance_tracker.theme.AppColors
 
 @Composable
-fun OperationsScreen(modifier: Modifier = Modifier) {
+fun MainScreen(modifier: Modifier = Modifier) {
     Column(
         modifier = modifier
             .fillMaxSize()
-            .background(AppColors.Purple700)
+            .background(AppColors.Purple200)
     ) {
         Text(
-            text = stringResource(R.string.operations_screen_text),
+            text = stringResource(R.string.main_screen_text),
             fontWeight = FontWeight.Bold,
             color = Color.White,
             modifier = modifier.align(Alignment.CenterHorizontally),
             textAlign = TextAlign.Center,
             fontSize = 24.sp
         )
-        TransactionsListScreen()
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun OperationsScreenPreview() {
-    OperationsScreen()
+fun MainScreenPreview() {
+    MainScreen()
 }
