+++ app/src/main/java/com/finance_tracker/finance_tracker/screens/transactions/TransactionsList.kt
@@ -1,4 +1,4 @@
-package com.finance_tracker.finance_tracker.sreens.transactions
+package com.finance_tracker.finance_tracker.screens.transactions
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
